Extract LS context propagation into configurable propagator class
