+++ lib/lightstep/tracer.rb
@@ -5,6 +5,7 @@ require 'opentracing'
 
 require 'lightstep/span'
 require 'lightstep/reporter'
+require 'lightstep/propagation/lightstep_propagator'
 require 'lightstep/transport/http_json'
 require 'lightstep/transport/nil'
 require 'lightstep/transport/callback'
@@ -14,6 +15,11 @@ module LightStep
     class Error < LightStep::Error; end
     class ConfigurationError < LightStep::Tracer::Error; end
 
+    DEFAULT_MAX_LOG_RECORDS = 1000
+    MIN_MAX_LOG_RECORDS = 1
+    DEFAULT_MAX_SPAN_RECORDS = 1000
+    MIN_MAX_SPAN_RECORDS = 1
+
     attr_reader :access_token, :guid
 
     # Initialize a new tracer. Either an access_token or a transport must be
@@ -22,10 +28,19 @@ module LightStep
     # @param access_token [String] The project access token when pushing to LightStep
     # @param transport [LightStep::Transport] How the data should be transported
     # @param tags [Hash] Tracer-level tags
+    # @param propagator [Propagator] Propagator to pass context over process boundaries
     # @return LightStep::Tracer
     # @raise LightStep::ConfigurationError if the group name or access token is not a valid string.
-    def initialize(component_name:, access_token: nil, transport: nil, tags: {})
-      configure(component_name: component_name, access_token: access_token, transport: transport, tags: tags)
+    def initialize(component_name:,
+                   access_token: nil,
+                   transport: nil,
+                   tags: {},
+                   propagator: Propagation::LightStepPropagator.new)
+      configure(component_name: component_name,
+                access_token: access_token,
+                transport: transport,
+                tags: tags,
+                propagator: propagator)
     end
 
     def max_log_records
@@ -172,22 +187,14 @@ module LightStep
       end
     end
 
+
     # Inject a SpanContext into the given carrier
     #
     # @param spancontext [SpanContext]
     # @param format [OpenTracing::FORMAT_TEXT_MAP, OpenTracing::FORMAT_BINARY]
     # @param carrier [Carrier] A carrier object of the type dictated by the specified `format`
     def inject(span_context, format, carrier)
-      case format
-      when OpenTracing::FORMAT_TEXT_MAP
-        inject_to_text_map(span_context, carrier)
-      when OpenTracing::FORMAT_BINARY
-        warn 'Binary inject format not yet implemented'
-      when OpenTracing::FORMAT_RACK
-        inject_to_rack(span_context, carrier)
-      else
-        warn 'Unknown inject format'
-      end
+      @propagator.inject(span_context, format, carrier)
     end
 
     # Extract a SpanContext from a carrier
@@ -195,18 +202,7 @@ module LightStep
     # @param carrier [Carrier] A carrier object of the type dictated by the specified `format`
     # @return [SpanContext] the extracted SpanContext or nil if none could be found
     def extract(format, carrier)
-      case format
-      when OpenTracing::FORMAT_TEXT_MAP
-        extract_from_text_map(carrier)
-      when OpenTracing::FORMAT_BINARY
-        warn 'Binary join format not yet implemented'
-        nil
-      when OpenTracing::FORMAT_RACK
-        extract_from_rack(carrier)
-      else
-        warn 'Unknown join format'
-        nil
-      end
+      @propagator.extract(format, carrier)
     end
 
     # @return true if the tracer is enabled
@@ -243,7 +239,7 @@ module LightStep
 
     protected
 
-    def configure(component_name:, access_token: nil, transport: nil, tags: {})
+    def configure(component_name:, access_token: nil, transport: nil, tags: {}, propagator: nil)
       raise ConfigurationError, "component_name must be a string" unless component_name.is_a?(String)
       raise ConfigurationError, "component_name cannot be blank"  if component_name.empty?
 
@@ -254,6 +250,9 @@ module LightStep
       raise ConfigurationError, "you must provide an access token or a transport" if transport.nil?
       raise ConfigurationError, "#{transport} is not a LightStep transport class" if !(LightStep::Transport::Base === transport)
 
+      #@todo: check propagator type
+      @propagator = propagator
+
       @guid = LightStep.guid
 
       @reporter = LightStep::Reporter.new(
@@ -263,76 +262,6 @@ module LightStep
         component_name: component_name,
         tags: tags
       )
-    end
-
-    private
-
-    CARRIER_TRACER_STATE_PREFIX = 'ot-tracer-'.freeze
-    CARRIER_BAGGAGE_PREFIX = 'ot-baggage-'.freeze
-
-    CARRIER_SPAN_ID = (CARRIER_TRACER_STATE_PREFIX + 'spanid').freeze
-    CARRIER_TRACE_ID = (CARRIER_TRACER_STATE_PREFIX + 'traceid').freeze
-    CARRIER_SAMPLED = (CARRIER_TRACER_STATE_PREFIX + 'sampled').freeze
-
-    DEFAULT_MAX_LOG_RECORDS = 1000
-    MIN_MAX_LOG_RECORDS = 1
-    DEFAULT_MAX_SPAN_RECORDS = 1000
-    MIN_MAX_SPAN_RECORDS = 1
-
-    def inject_to_text_map(span_context, carrier)
-      carrier[CARRIER_SPAN_ID] = span_context.id
-      carrier[CARRIER_TRACE_ID] = span_context.trace_id unless span_context.trace_id.nil?
-      carrier[CARRIER_SAMPLED] = 'true'
-
-      span_context.baggage.each do |key, value|
-        carrier[CARRIER_BAGGAGE_PREFIX + key] = value
-      end
-    end
-
-    def extract_from_text_map(carrier)
-      # If the carrier does not have both the span_id and trace_id key
-      # skip the processing and just return a normal span
-      if !carrier.has_key?(CARRIER_SPAN_ID) || !carrier.has_key?(CARRIER_TRACE_ID)
-        return nil
-      end
-
-      baggage = carrier.reduce({}) do |baggage, tuple|
-        key, value = tuple
-        if key.start_with?(CARRIER_BAGGAGE_PREFIX)
-          plain_key = key.to_s[CARRIER_BAGGAGE_PREFIX.length..key.to_s.length]
-          baggage[plain_key] = value
-        end
-        baggage
-      end
-      SpanContext.new(
-        id: carrier[CARRIER_SPAN_ID],
-        trace_id: carrier[CARRIER_TRACE_ID],
-        baggage: baggage,
-      )
-    end
-
-    def inject_to_rack(span_context, carrier)
-      carrier[CARRIER_SPAN_ID] = span_context.id
-      carrier[CARRIER_TRACE_ID] = span_context.trace_id unless span_context.trace_id.nil?
-      carrier[CARRIER_SAMPLED] = 'true'
-
-      span_context.baggage.each do |key, value|
-        if key =~ /[^A-Za-z0-9\-_]/
-          # TODO: log the error internally
-          next
-        end
-        carrier[CARRIER_BAGGAGE_PREFIX + key] = value
-      end
-    end
-
-    def extract_from_rack(env)
-      extract_from_text_map(env.reduce({}){|memo, tuple|
-        raw_header, value = tuple
-        header = raw_header.to_s.gsub(/^HTTP_/, '').tr('_', '-').downcase
-
-        memo[header] = value if header.start_with?(CARRIER_TRACER_STATE_PREFIX, CARRIER_BAGGAGE_PREFIX)
-        memo
-      })
     end
   end
 end
